Use Metal shared events for command buffer sync
Add a signaled value to MetalCommandBuffer and create a MTLSharedEvent
on MetalDevice. Command buffers encodeSignalEvent before commit and
waiters use waitUntilSignaledValue on the shared event instead of
blocking on the MTLCommandBuffer. Also cast push constant size to
uint32_t in VulkanCommandBuffer to match the Vulkan API.

+++ src/Vulkan/VulkanCommandBuffer.cpp
@@ -314,7 +314,7 @@ void VulkanCommandBuffer::setParameterBlock(const std::shared_ptr<const Paramete
 void VulkanCommandBuffer::setPushConstants(const void* data, size_t size)
 {
     assert(m_boundPipeline != nullptr);
-    m_vkCommandBuffer.pushConstants(m_boundPipeline->pipelineLayout(), vk::ShaderStageFlagBits::eVertex | vk::ShaderStageFlagBits::eFragment, 0, size, data);
+    m_vkCommandBuffer.pushConstants(m_boundPipeline->pipelineLayout(), vk::ShaderStageFlagBits::eVertex | vk::ShaderStageFlagBits::eFragment, 0, (uint32_t)size, data);
 }
 
 void VulkanCommandBuffer::drawVertices(uint32_t start, uint32_t count)
